backend/pyproject.toml
[project]
name = "nexus-core"
version = "0.1.0"
description = "Nexus Core - Multi-Persona AI Chatbot Platform Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.7.0",
    "google-genai>=1.63.0",
    "openai>=2.21.0",
    "python-multipart>=0.0.22",
    "aiofiles>=24.1.0",
    "greenlet>=3.3.1",
    "pyjwt[crypto]>=2.9.0",
    "pgvector>=0.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]
